requirements.txt
# Core ML & Data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
scipy>=1.11.0

# Deep Learning (optional, for advanced models)
# torch>=2.1.0

# Network traffic parsing
scapy>=2.5.0
dpkt>=1.9.8

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Data storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Monitoring & Logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Utilities
click>=8.1.0
rich>=13.7.0
joblib>=1.3.0
